+++ src/Views/Classic/Styles/Vertical_Tabs.php
@@ -216,31 +216,4 @@ class Vertical_Tabs extends Styling_Base {
 		</div>
 		<?php
 	}
-
-	/**
-	 * Output the HTML-code for the settings.
-	 *
-	 * @param Tab $tab The tab to show.
-	 *
-	 * @return void
-	 */
-	public function show_content( Tab $tab ): void {
-		// show the tab description.
-		if ( ! empty( $tab->get_description() ) ) {
-			echo wp_kses_post( $tab->get_description() );
-		}
-
-		?>
-		<form method="POST" action="<?php echo esc_url( get_admin_url() ); ?>options.php">
-			<?php
-			if ( 'options-general.php' !== $this->settings_obj->get_menu_parent_slug() ) {
-				settings_errors(); }
-			settings_fields( $tab->get_name() );
-			do_settings_sections( $tab->get_name() );
-			if ( ! $tab->is_save_hidden() ) {
-				submit_button(); }
-			?>
-		</form>
-		<?php
-	}
 }